+++ srcs/backend/src/block/block.service.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from 'src/prisma/prisma.service'; // Assurez-vous d'utiliser le chemin correct
 import { Block, Prisma } from '@prisma/client';
 
 @Injectable()
@@ -10,8 +10,16 @@ export class BlockService {
     return this.prisma.block.create({ data });
   }
 
-  async getBlockById(id: number): Promise<Block | null> {
-    return this.prisma.block.findUnique({ where: { id } });
+  async getBlockById(id: number): Promise<Block> {
+    const block = await this.prisma.block.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!block) {
+      throw new NotFoundException('Block not found');
+    }
+    return block;
   }
 
   async getAllBlocks(): Promise<Block[]> {
@@ -19,6 +27,10 @@ export class BlockService {
   }
 
   async updateBlock(id: number, data: Prisma.BlockUpdateInput): Promise<Block | null> {
+    const existingBlock = await this.prisma.block.findUnique({ where: { id } });
+    if (!existingBlock) {
+      throw new NotFoundException(`Block with ID ${id} not found`);
+    }
     return this.prisma.block.update({
       where: { id },
       data,
@@ -26,6 +38,10 @@ export class BlockService {
   }
 
   async deleteBlock(id: number): Promise<Block | null> {
+    const existingBlock = await this.prisma.block.findUnique({ where: { id } });
+    if (!existingBlock) {
+      throw new NotFoundException(`Block with ID ${id} not found`);
+    }
     return this.prisma.block.delete({ where: { id } });
   }
 }
