+++ srcs/backend/src/friendrequest/friendrequest.service.ts
@@ -1,5 +1,5 @@
-import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/prisma/prisma.service';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaService } from 'src/prisma/prisma.service'; // Assurez-vous d'utiliser le chemin correct
 import { FriendRequest, Prisma } from '@prisma/client';
 
 @Injectable()
@@ -10,8 +10,16 @@ export class FriendRequestService {
     return this.prisma.friendRequest.create({ data });
   }
 
-  async getFriendRequestById(id: number): Promise<FriendRequest | null> {
-    return this.prisma.friendRequest.findUnique({ where: { id } });
+  async getFriendRequestById(id: number): Promise<FriendRequest> {
+    const request = await this.prisma.friendRequest.findUnique({
+      where: {
+        id,
+      },
+    });
+    if (!request) {
+      throw new NotFoundException('Friend request not found');
+    }
+    return request;
   }
 
   async getAllFriendRequests(): Promise<FriendRequest[]> {
@@ -19,6 +27,10 @@ export class FriendRequestService {
   }
 
   async updateFriendRequest(id: number, data: Prisma.FriendRequestUpdateInput): Promise<FriendRequest | null> {
+    const existingRequest = await this.prisma.friendRequest.findUnique({ where: { id } });
+    if (!existingRequest) {
+      throw new NotFoundException(`Friend request with ID ${id} not found`);
+    }
     return this.prisma.friendRequest.update({
       where: { id },
       data,
@@ -26,6 +38,10 @@ export class FriendRequestService {
   }
 
   async deleteFriendRequest(id: number): Promise<FriendRequest | null> {
+    const existingRequest = await this.prisma.friendRequest.findUnique({ where: { id } });
+    if (!existingRequest) {
+      throw new NotFoundException(`Friend request with ID ${id} not found`);
+    }
     return this.prisma.friendRequest.delete({ where: { id } });
   }
 }
